fix: sign-in screen loading issue // TODO: Logic for sign in
