+++ package.json
@@ -36,6 +36,7 @@
     "expo-system-ui": "~4.0.9",
     "expo-web-browser": "~14.0.2",
     "formik": "^2.4.6",
+    "lucide-react-native": "^0.487.0",
     "nativewind": "^4.1.23",
     "react": "18.3.1",
     "react-dom": "18.3.1",
@@ -49,7 +50,7 @@
     "react-native-reanimated": "~3.16.1",
     "react-native-safe-area-context": "4.12.0",
     "react-native-screens": "~4.4.0",
-    "react-native-svg": "^15.8.0",
+    "react-native-svg": "15.8.0",
     "react-native-toast-message": "^2.2.1",
     "react-native-utils-scale": "^2.6.0",
     "react-native-uuid": "^2.0.3",
