+++ svgs/LogoIcon.tsx
@@ -1,0 +1,21 @@
+import * as React from "react"
+import Svg, { Path } from "react-native-svg"
+const LogoIcon = (props: any) => (
+	<Svg
+		xmlns="http://www.w3.org/2000/svg"
+		width={58}
+		height={53}
+		fill="none"
+		{...props}
+	>
+		<Path
+			fill="#5ED6D5"
+			d="M28.536 0c-1.952 0-3.917.397-5.485 1.264L2.627 12.558c-.784.433-1.42.948-1.88 1.572A3.796 3.796 0 0 0 0 16.375c0 .838.288 1.619.748 2.243.46.625 1.095 1.14 1.88 1.573l2.243 1.238-2.244 1.238c-.784.433-1.4.947-1.86 1.572-.46.626-.767 1.406-.767 2.244 0 .838.307 1.618.767 2.243.46.625 1.076 1.14 1.86 1.573l2.244 1.238-2.244 1.237c-.784.434-1.4.949-1.86 1.574C.307 34.972 0 35.752 0 36.59c0 .838.307 1.618.767 2.243.46.625 1.076 1.14 1.86 1.573L23.051 51.7c3.136 1.733 7.795 1.733 10.931 0l20.462-11.294c.784-.433 1.4-.948 1.86-1.573.46-.625.767-1.405.767-2.243 0-.838-.307-1.618-.767-2.243-.46-.625-1.076-1.14-1.86-1.574l-2.263-1.237 2.263-1.238c.784-.433 1.4-.947 1.86-1.573.46-.625.767-1.405.767-2.243 0-.838-.307-1.618-.767-2.244-.46-.625-1.076-1.14-1.86-1.572l-2.263-1.238 2.244-1.238c.784-.433 1.42-.948 1.88-1.573.459-.624.747-1.405.747-2.243 0-.838-.288-1.62-.748-2.243-.46-.625-1.095-1.14-1.88-1.573L34.002 1.264C32.433.397 30.487 0 28.536 0ZM6.885 22.563l16.166 8.921c3.136 1.733 7.795 1.733 10.931 0l4.373-2.423 11.832-6.498 3.663 2.01c.65.36 1.104.76 1.361 1.11.257.35.326.6.326.799 0 .198-.069.45-.326.8-.257.35-.712.749-1.361 1.108L33.407 39.684c-.975.539-2.228.897-3.548 1.032-2.2.224-4.588-.134-6.214-1.032L3.203 28.39c-.65-.36-1.105-.759-1.362-1.109-.257-.35-.307-.601-.307-.8 0-.197.05-.448.307-.798.257-.35.712-.75 1.362-1.11l3.682-2.01Zm0 10.108 16.166 8.922c1.96 1.082 4.517 1.508 6.923 1.237 1.444-.162 2.832-.588 4.008-1.237l16.205-8.922 3.663 2.01c.65.36 1.104.76 1.361 1.11.257.35.326.6.326.799 0 .198-.069.45-.326.8-.257.35-.712.749-1.361 1.108L33.407 49.792c-2.6 1.437-7.161 1.437-9.761 0L3.202 38.498c-.65-.359-1.105-.759-1.362-1.109-.257-.35-.307-.6-.307-.8 0-.197.05-.448.307-.798.257-.35.712-.75 1.362-1.11l3.682-2.01Z"
+		/>
+		<Path
+			fill="#fff"
+			d="M28.642 22.248c-.987.203-2.026.145-3.117-.145-1.09-.29-2.13-.812-3.117-1.595-1.454-1.16-2.234-2.378-2.39-3.626-.13-1.218.416-2.291 1.663-3.248l3.195 2.407c-.442.406-.675.812-.675 1.247 0 .436.233.842.7 1.19.416.319.858.464 1.326.493.467 0 .883-.116 1.272-.406.338-.261.52-.551.52-.87 0-.32-.078-.668-.26-1.016s-.468-.783-.831-1.334c-.52-.783-.91-1.45-1.169-2.001a2.57 2.57 0 0 1-.13-1.711c.156-.58.623-1.16 1.403-1.683a5.403 5.403 0 0 1 2.39-.899c.882-.116 1.791-.029 2.727.203.935.232 1.843.667 2.727 1.218 1.454.928 2.337 1.944 2.623 2.988.286 1.073-.052 2.088-1.039 3.045l-3.584-2.436c.337-.348.493-.696.467-1.044-.026-.348-.26-.667-.675-.957a2.482 2.482 0 0 0-1.169-.406c-.415-.03-.78.087-1.143.348-.285.203-.441.435-.441.725s.078.58.26.87c.181.319.467.725.857 1.305a17.257 17.257 0 0 1 1.324 2.088c.286.61.39 1.248.286 1.973-.104.725-.597 1.42-1.455 2.088-.7.638-1.558 1.015-2.545 1.19Z"
+		/>
+	</Svg>
+)
+export default LogoIcon
